requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
apscheduler==3.10.4
aiosqlite==0.19.0
pymorphy3>=2.0.0
pymorphy3-dicts-ru>=2.4.0
gspread==5.12.0
google-auth==2.25.2

